+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - Users may still manually run from source for future builds
   - Binaries will no longer be provided on future release
 - Switch boot.efi model patch to iMac18,1
+- Resolve pre-Force Touch Trackpad support in Ventura
 
 ## 0.4.5
 - Fix AutoPatcher.pkg download on releases